Runtime nmodl flags replace build nmodl flags
Implementing the logic to sanitize the merge of default and
runtime nmodl flags was a little too much for a makefile.
We let the user add the runtime option at his own risk. We
inform him on what were the default flags to let him make an
easy merge with what he wants.

- Add warning and info to be explicit