Define NrnCoreTransferEvents in global namespace.
A pointer to this type is passed across the NEURON-CoreNEURON boundary
by the nrn2core_transfer_tqueue callback, and it is undefined behaviour
to call a function via a pointer to a different function type. This is
what was happening before due to a mismatch of NrnCoreTransferEvents and
coreneuron::NrnCoreTransferEvents.