Improve GitLab pipeline.
Streamline the pipeline a little by using a single build of NMODL with
GCC instead of building it with both Intel and NVHPC. Also allocate more
resources for build jobs to avoid too many of them landing on the same
nodes and killing the filesystem. Extend an existing fix for NVHPC to
avoid building py-{cython,numpy} with Intel.

Temporary changes to be reverted:
 - enable ccache debugging
 - use an alternative branch of gitlab-pipelines